Extracts batch action module into its own file
